+++ handler.go
@@ -23,6 +23,9 @@ type handler struct {
 }
 
 func newHandler(do reflect.Value) *handler {
+	if do.Type().NumOut() <= 0 {
+		panic(fmt.Sprintf("handling function should be non-void function but got: %s", do.Type()))
+	}
 	return &handler{
 		do:            do,
 		wildcardIndex: -1,
@@ -121,6 +124,9 @@ func (h *handler) fillByCachedUserContexts(contexts []*context.UserContext, reqU
 		}
 		if userContext.ExpectHeader() {
 			basicChain = append(basicChain, newHeaderFiller(userContext.HeaderParams, req.Header))
+		}
+		if userContext.ExpectHTTP() {
+			basicChain = append(basicChain, newHTTPFiller(userContext.HttpParams, req))
 		}
 		ctx := reflect.New(userContext.ContextType)
 		for _, filler := range basicChain {