+++ handling_test.go
@@ -1,16 +1,66 @@
 package rocket_test
 
 import (
+	"crypto/tls"
+	"crypto/x509"
+	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
 	"github.com/dannypsnl/rocket"
-	"github.com/dannypsnl/rocket/fairing"
 
-	"github.com/dannypsnl/assert"
 	"github.com/gavv/httpexpect"
+	"github.com/stretchr/testify/assert"
+	"golang.org/x/net/http2"
 )
+
+func TestHTTPsAndHTTP2(t *testing.T) {
+	rk := rocket.Ignite(":8082").
+		Mount(rocket.Get("/", func() string { return "home" }))
+	ts := httptest.NewUnstartedServer(rk)
+	ts.TLS = &tls.Config{
+		CipherSuites: []uint16{tls.TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256},
+		NextProtos:   []string{http2.NextProtoTLS},
+	}
+	ts.StartTLS()
+	defer ts.Close()
+
+	// Create a pool with the server certificate since it is not signed
+	// by a known CA
+	caCert := ts.Certificate()
+	caCertPool := x509.NewCertPool()
+	caCertPool.AppendCertsFromPEM(caCert.RawTBSCertificate)
+
+	// Create TLS configuration with the certificate of the server
+	tlsConfig := &tls.Config{
+		InsecureSkipVerify: true,
+		RootCAs:            caCertPool,
+	}
+	client := &http.Client{}
+	client.Transport = &http2.Transport{
+		TLSClientConfig: tlsConfig,
+	}
+
+	resp, err := client.Get(ts.URL)
+	if err != nil {
+		t.Fatalf("failed at GET, error: %s", err)
+	}
+	if resp.StatusCode != 200 {
+		t.Error("request fail")
+	}
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		t.Fatalf("failed at read response body, error: %s", err)
+	}
+	resp.Body.Close()
+	if string(body) != "home" {
+		t.Error("response body is wrong")
+	}
+	if resp.Proto != "HTTP/2.0" {
+		t.Error("protocol should be HTTP/2")
+	}
+}
 
 var (
 	forTestHandler = rocket.Get("/", func() string { return "" })
@@ -30,7 +80,7 @@ func TestOptionsMethod(t *testing.T) {
 }
 
 type Recorder struct {
-	fairing.Fairing
+	rocket.Fairing
 
 	RecordRequestURL []string
 }
@@ -41,8 +91,6 @@ func (r *Recorder) OnRequest(req *http.Request) *http.Request {
 }
 
 func TestRecorder(t *testing.T) {
-	assert := assert.NewTester(t)
-
 	recorder := &Recorder{
 		RecordRequestURL: make([]string, 0),
 	}
@@ -58,7 +106,7 @@ func TestRecorder(t *testing.T) {
 	e.GET("/").
 		Expect().Status(http.StatusOK)
 
-	assert.Eq(recorder.RecordRequestURL[0], "/")
+	assert.Equal(t, "/", recorder.RecordRequestURL[0])
 }
 
 type AccessCookie struct {
@@ -100,4 +148,23 @@ func TestGetHeaderByUserDefinedContext(t *testing.T) {
 	e.GET("/").WithHeader("Authorization", "Bear jwt.token.lalala").
 		Expect().Status(http.StatusOK).
 		Body().Equal("Bear jwt.token.lalala")
+}
+
+type RequestContext struct {
+	Request *http.Request `http:"request"`
+}
+
+func TestHTTPInvalidResourceShouldBeRejected(t *testing.T) {
+	rk := rocket.Ignite("").
+		Mount(rocket.Get("/path", func(req *RequestContext) string {
+			return req.Request.URL.Path
+		}))
+
+	ts := httptest.NewServer(rk)
+	defer ts.Close()
+	e := httpexpect.New(t, ts.URL)
+
+	e.GET("/path").
+		Expect().Status(http.StatusOK).
+		Body().Equal("/path")
 }